network: Change subscribeTo method signature
It was noticed that subscription doesn't need to have many routing keys
associated. This patch changes the subscribeTo method, to receive only
one routing key instead of a list of these keys. Also, change the "on"
Client method, due to the non data associated events are now returning
responses with the respective requests.