client: Fix error catching in callback function
It was noticed that if occur an error when unsubscribing the consumer
before processing the message data, an error will be thrown, neither
resolving or rejecting the promise. To fix this behavior, the
unsubscribing method was changed to not be blocking anymore. Also, add a
catch statement to handle errors when publishing message, after
subscribing a consumer.