client: Add session to 'data' event exchange name
The user session mechanism was added to prevent the user from receiving
messages that belongs to other users. Instead of just subscribing to
'data' exchange, it's necessary to create a session ID with the provided
token and subscribe to 'data.<session-id>.published`.

In order to facilitate this process to the users, this patch brings this
session creation to the SDK. As a result, there's no breaking change in
the `on` operation but the user must now configure the SDK with `HTTP`
properties.
                                                         pick